"ads1_remedial_3_a1_1: 1/7 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 0/1 "

+++ ADS-Assignments/remedial/Remedial-3/Solution.java
@@ -8,7 +8,7 @@ public class Solution {
 		scan.nextLine();
 		for(int i =0;i<range;i++) {
 			String[] temp = scan.nextLine().split(",");
-			Student student = new Student(Integer.parseInt(temp[0]),temp[1],Integer.parseInt(temp[2]));
+			Student student = new Student(Integer.parseInt(temp[0]),temp[1],Double.parseDouble(temp[2]));
 			sort.add(student);
 
 		}
